Migrationしなおし＆エラーとり

+++ Models/InferenceAssumption.cs
@@ -21,7 +21,7 @@ public class InferenceAssumption
   /// <summary>
   /// 論理式構成ID
   /// </summary>
-  public int FormulaStructId { get; set; }
+  public long FormulaStructId { get; set; }
   public List<InferenceFormulaStructArgumentMapping> FormulaStructArgumentMappings { get; set; } = new();
   /// <summary>
   /// 論理式構成-引数マッピング
