+++ Models/InferenceConclusion.cs
@@ -17,7 +17,7 @@ public class InferenceConclusion
   /// <summary>
   /// 論理式構成ID
   /// </summary>
-  public int FormulaStructId { get; set; }
+  public long FormulaStructId { get; set; }
   public List<InferenceFormulaStructArgumentMapping> FormulaStructArgumentMappings { get; set; } = new();
   /// <summary>
   /// 論理式構成-引数マッピング
